[#8] custom navigation bar frame 수정, VCRepresentable 추가, SwiftUI 뷰와 연결
