--- Project/Reazy/Views/PDF/OriginalView.swift
@@ -1,28 +1,0 @@
-//
-//  OriginalView.swift
-//  Reazy
-//
-//  Created by 유지수 on 10/14/24.
-//
-
-import SwiftUI
-
-// MARK: - 무니꺼 : 원문 모드 뷰
-struct OriginalView: View {
-  var body: some View {
-    VStack(spacing: 0) {
-      Spacer()
-      HStack() {
-        Spacer()
-        Text("원문 모드")
-        Spacer()
-      }
-      Spacer()
-    }
-    .background(.gray)
-  }
-}
-
-#Preview {
-  OriginalView()
-}
